Add files via upload
I cannot upload the software as github doesnt allow upload of files above 100. hence uploading the program. 
Note: Please do not use the program for illegal purposes. The program is open sourced only to learn the significance of beautifulsoup  library in python and modifications, suggestions  for the codebase are always welcome if its genuine one.